requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coverage==5.2.1
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.1
django-debug-toolbar==3.2
django-debug-toolbar-force==0.1.8
django-filter==2.3.0
django-nine==0.2.4
djangorestframework==3.11.1
idna==2.10
Markdown==3.3.3
mdown==2020.12.3
mkdir==2020.12.3
oauthlib==3.1.0
orderdict==2020.12.3
psycopg2==2.8.5
public==2020.12.3
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
readme-md==2019.1.28
requests==2.24.0
requests-oauthlib==1.3.0
setupcfg==2020.12.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
values==2020.12.3
write==2020.12.3
